WebStorage/tests/approved/storage_local_removeitem.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>storage_local_removeItem</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>

		if (('localStorage' in window) && window.localStorage !== null){
			localStorage.clear();
			localStorage.setItem("name", "user1");

			if (localStorage.getItem("name") == "user1")
			{
				try {
					localStorage.removeItem("name");
					localStorage.removeItem("unknown");
				} catch (e) {
					test(function() {
						assert_unreached("removeItem method is failed.");
					}, "removeItem test")
				}

				test(function() {
					assert_equals(localStorage.getItem("name"), null, "localStorage.getItem('name')")
				}, "result test of removeItem method")

			} else {
				test(function() {
					assert_unreached("'name' item is not exist.");
				}, "item existence test")
			}
		} else {
			test(function()
			{
				assert_unreached("localStorage is not supported.");
			}, "localStorage support test")
		}	
	</script>

 </body>
</html>
